Sample code for controlling the pir sensor
Code I put together for the sensor. The led in the code is external to
the pi and the pir seonsor, I can set you with some parts if you want.
it is not important for the sensor. Just comment out the lines with
GIPO. Also make sure, on the pir sensor, that the potentiometers
controlling sensitivity is turned to max and the time delay to min.